@@ -18,38 +18,64 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+11
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
+  flutter: ">=2.5.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.13.0
-  xml: ^5.0.2
-  google_maps_flutter: ^2.0.1
-  maps_launcher: ^2.0.0
-  shared_preferences: ^2.0.5
-  url_launcher: ^6.0.3
-  flutter_local_notifications: ^5.0.0+1
-  background_fetch: ^0.7.2
+
+
+  http: ^0.13.3
+  xml: ^5.3.0
+  google_maps_flutter: ^2.0.10
+  maps_launcher: ^2.0.1
+  shared_preferences: ^2.0.8
+  url_launcher: ^6.0.12
+  flutter_local_notifications: ^8.2.0
+  background_fetch: ^1.0.1
   searchable_dropdown: ^1.1.3
-  dropdown_search: #^0.6.1 #- can't use until they fix keyboard overlapping bottom sheet search bar
-    path: ./local-plugins/dropdown_search-0.6.1-local
-  sqflite: ^2.0.0+3
+  dropdown_search: ^1.0.4 #^0.6.1 #- can't use until they fix keyboard overlapping bottom sheet search bar
+   # path: ./local-plugins/dropdown_search-0.6.1-local
+  sqflite: ^2.0.0+4
   introduction_screen: ^2.1.0
   intl: ^0.17.0
-  flutter_keyboard_visibility: ^5.0.1
+  flutter_keyboard_visibility: ^5.0.3
   pull_to_refresh: ^2.0.0
   provider: ^5.0.0
-  firebase_messaging: ^10.0.3
+  firebase_messaging: ^10.0.7
+  flutter_layout_grid: ^1.0.3
+  flutter_native_splash: ^1.2.4
+  flutter_launcher_icons: ^0.9.2
 
-  cupertino_icons: ^0.1.3
+  device_preview: ^0.8.0
+
+  cupertino_icons: ^1.0.3
 
 dev_dependencies:
   test: any
   flutter_driver:
     sdk: flutter
-  mockito: ^5.0.7
-  build_runner: ^2.0.1
+  mockito: ^5.0.16
+  build_runner: ^2.1.2
+  flutter_lints: ^1.0.4
+
+
+# flutter pub run flutter_native_splash:create
+flutter_native_splash:
+  color: '#d96d19'
+  fullscreen: true
+  image: 'assets/splash/Perth Fuel Prices Splash Screen.png'
+
+
+# flutter pub run flutter_launcher_icons:main
+flutter_icons:
+  android: "ic_launcher"
+  ios: true
+  image_path: "assets/icon/Perth Fuel Prices App Icon.png"
+  remov_alpha_ios: true
+
+
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -65,6 +91,7 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/images/
+    - assets/icon/
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
